refactor: move ListTable to component folder
