+++ src/Views/Components/index.js
@@ -11,3 +11,4 @@ export {default as Button} from './Button';
 export {default as Input} from './Input';
 export {default as PeriodSelector} from './PeriodSelector';
 export {default as Modal} from './Modal';
+export {default as ListTable} from './ListTable';
